requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0
pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.6
numpy>=1.24.0
matplotlib>=3.7.0